+++ src/main/java/roomescape/controller/ReservationApiController.java
@@ -3,9 +3,9 @@ package roomescape.controller;
 import jakarta.servlet.http.HttpServletResponse;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.*;
-import roomescape.dto.ReservationRequest;
-import roomescape.dto.ReservationCreateResponse;
-import roomescape.dto.ReservationsResponse;
+import roomescape.dto.request.ReservationRequest;
+import roomescape.dto.response.ReservationCreateResponse;
+import roomescape.dto.response.ReservationsResponse;
 import roomescape.service.ReservationService;
