+++ src/main/java/roomescape/dto/response/ReservationsResponse.java
@@ -1,4 +1,4 @@
-package roomescape.dto;
+package roomescape.dto.response;
 
 import roomescape.domain.Reservation;
